The purpose of this branch is to build ios_sim_arm64 binaries for both python and numpy. This branch is based on a pull-request that supports arm64-simulator as this functionality doesn't yet exist on the main line of kivy-ios.
The python code uses the patches and setup code from Krishna's origin
python-3.9.1 branch. These are essentially "drop-replaced" in the
arm64-simulator aware __init__.py script. The size of the patch used
used is a bit alarming - Krishna's notes say it was brought over from
the `beeware` project. See the BFLY.md readme.

The numpy code build is a bit trickier - the head of the python-3.9.1
branch was _not_ used to generate our current production libraries
(slight symbol difference see `_gemms_`).
Here I've opted to use a slightly more recent main line branch commit
of 1.20.2:

   commit e38cd58b16aed04aae6701e4f2fb4b1dbf854f44
   Merge: 6b245b1 84595ab
   Author: Andre Miras <AndreMiras@users.noreply.github.com>
   Date:   Sun Apr 4 20:25:20 2021 +0200

       Merge pull request #604 from misl6/feat/upgrade-numpy

       Update numpy to version 1.20.2

This is not 1.20.1, but should be compatible. It also addresses the
Accelerate framework issue.

Finally - we had to fix up the hostpython3 environment as this failed
without patching.  Here we just grabbed the patch used by pyenv for
building and used that.